[CDEC-484] Reunite orphans from Pos.Chain.Ssc.Toss.Trans
