+++ chain/src/Pos/Chain/Ssc/Toss.hs
@@ -4,7 +4,6 @@ module Pos.Chain.Ssc.Toss
        , module Pos.Chain.Ssc.Toss.Class
        , module Pos.Chain.Ssc.Toss.Logic
        , module Pos.Chain.Ssc.Toss.Pure
-       , module Pos.Chain.Ssc.Toss.Trans
        , module Pos.Chain.Ssc.Toss.Types
        ) where
 
@@ -12,6 +11,4 @@ import           Pos.Chain.Ssc.Toss.Base
 import           Pos.Chain.Ssc.Toss.Class
 import           Pos.Chain.Ssc.Toss.Logic
 import           Pos.Chain.Ssc.Toss.Pure
-import           Pos.Chain.Ssc.Toss.Trans
 import           Pos.Chain.Ssc.Toss.Types
-
