+++ chain/src/Pos/Chain/Ssc/Toss/Types.hs
@@ -10,20 +10,30 @@ module Pos.Chain.Ssc.Toss.Types
        , tmOpenings
        , tmShares
        , tmCertificates
+
+       , TossT
+       , runTossT
+       , evalTossT
+       , execTossT
        ) where
 
-import           Control.Lens (makeLenses)
+import           Universum hiding (id)
+
+import           Control.Lens (at, makeLenses, (%=), (.=))
+import qualified Ether
 import qualified Formatting.Buildable as Buildable
-import           Universum
 
 import           Pos.Binary.Class (Cons (..), Field (..), deriveSimpleBi,
                      deriveSimpleBiCxt)
-import           Pos.Chain.Ssc.Base (isCommitmentId, isCommitmentIdx,
+import           Pos.Chain.Ssc.Base (deleteSignedCommitment,
+                     insertSignedCommitment, isCommitmentId, isCommitmentIdx,
                      isOpeningId, isOpeningIdx, isSharesId, isSharesIdx)
+import           Pos.Chain.Ssc.Toss.Class (MonadToss (..), MonadTossEnv (..),
+                     MonadTossRead (..))
 import           Pos.Core (HasProtocolConstants, LocalSlotIndex, SlotId)
 import           Pos.Core.Ssc (CommitmentsMap, OpeningsMap, SharesMap,
-                     VssCertificatesMap)
-import           Pos.Util.Util (cborError)
+                     VssCertificatesMap, insertVss)
+import           Pos.Util.Util (cborError, ether)
 
 -- | Tag corresponding to SSC data.
 data SscTag
@@ -38,12 +48,6 @@ instance Buildable SscTag where
     build OpeningMsg        = "opening"
     build SharesMsg         = "shares"
     build VssCertificateMsg = "VSS certificate"
-
-deriveSimpleBi ''SscTag [
-    Cons 'CommitmentMsg [],
-    Cons 'OpeningMsg [],
-    Cons 'SharesMsg [],
-    Cons 'VssCertificateMsg []]
 
 isGoodSlotForTag :: HasProtocolConstants => SscTag -> LocalSlotIndex -> Bool
 isGoodSlotForTag CommitmentMsg     = isCommitmentIdx
@@ -64,8 +68,6 @@ data TossModifier = TossModifier
     , _tmCertificates :: !VssCertificatesMap
     } deriving (Generic, Show, Eq)
 
-makeLenses ''TossModifier
-
 instance Semigroup TossModifier where
    (TossModifier leftComms leftOpens leftShares leftCerts)
      <> (TossModifier rightComms rightOpens rightShares rightCerts) =
@@ -80,6 +82,79 @@ instance Monoid TossModifier where
     mempty = TossModifier mempty mempty mempty mempty
     mappend = (<>)
 
+----------------------------------------------------------------------------
+-- Tranformer
+----------------------------------------------------------------------------
+
+-- | Monad transformer which stores TossModifier and implements
+-- writable MonadToss.
+--
+-- [WARNING] This transformer uses StateT and is intended for
+-- single-threaded usage only.
+type TossT = Ether.StateT' TossModifier
+
+----------------------------------------------------------------------------
+-- Runners
+----------------------------------------------------------------------------
+
+runTossT :: TossModifier -> TossT m a -> m (a, TossModifier)
+runTossT = flip Ether.runStateT
+
+evalTossT :: Monad m => TossModifier -> TossT m a -> m a
+evalTossT = flip Ether.evalStateT
+
+execTossT :: Monad m => TossModifier -> TossT m a -> m TossModifier
+execTossT = flip Ether.execStateT
+
+----------------------------------------------------------------------------
+-- MonadToss
+----------------------------------------------------------------------------
+
+makeLenses ''TossModifier
+
+instance MonadTossRead m =>
+         MonadTossRead (TossT m) where
+    getCommitments = ether $ (<>) <$> use tmCommitments <*> getCommitments
+    getOpenings = ether $ (<>) <$> use tmOpenings <*> getOpenings
+    getShares = ether $ (<>) <$> use tmShares <*> getShares
+    getVssCertificates = ether $ (<>) <$> use tmCertificates <*> getVssCertificates
+    getStableCertificates = ether . getStableCertificates
+
+instance MonadTossEnv m =>
+         MonadTossEnv (TossT m) where
+    getRichmen = ether . getRichmen
+    getAdoptedBVData = ether getAdoptedBVData
+
+instance MonadToss m =>
+         MonadToss (TossT m) where
+    putCommitment signedComm =
+        ether $ tmCommitments %= insertSignedCommitment signedComm
+    putOpening id op =
+        ether $ tmOpenings . at id .= Just op
+    putShares id sh =
+        ether $ tmShares . at id .= Just sh
+    -- NB. 'insertVss' might delete some certs from the map, but it
+    -- shouldn't actually happen in practice because
+    -- 'checkCertificatesPayload' ensures that there are no clashes between
+    -- the certificates in blocks and certificates in the map
+    putCertificate cert =
+        ether $ tmCertificates %= fst . insertVss cert
+    delCommitment id =
+        ether $ tmCommitments %= deleteSignedCommitment id
+    delOpening id =
+        ether $ tmOpenings . at id .= Nothing
+    delShares id =
+        ether $ tmShares . at id .= Nothing
+    resetCO = ether $ do
+        tmCommitments .= mempty
+        tmOpenings .= mempty
+        tmCertificates .= mempty
+        resetCO
+    resetShares = ether $ do
+        tmShares .= mempty
+        resetShares
+    setEpochOrSlot = ether . setEpochOrSlot
+
 deriveSimpleBiCxt [t|()|] ''TossModifier [
     Cons 'TossModifier [
         Field [| _tmCommitments  :: CommitmentsMap     |],
@@ -87,3 +162,9 @@ deriveSimpleBiCxt [t|()|] ''TossModifier [
         Field [| _tmShares       :: SharesMap          |],
         Field [| _tmCertificates :: VssCertificatesMap |]
     ]]
+
+deriveSimpleBi ''SscTag [
+    Cons 'CommitmentMsg [],
+    Cons 'OpeningMsg [],
+    Cons 'SharesMsg [],
+    Cons 'VssCertificateMsg []]
